*** empty log message ***

+++ org/opensha/sha/earthquake/rupForecastImpl/WGCEP_UCERF_2_1/EqkRateModel2_ERF.java
@@ -204,7 +204,7 @@ public class EqkRateModel2_ERF extends EqkRupForecast {
 	private final static String REL_A_PRIORI_WT_PARAM_NAME = "Wt On A-Priori Rates";
 	private final static Double REL_A_PRIORI_WT_PARAM_MIN = new Double(Double.MIN_VALUE);
 	private final static Double REL_A_PRIORI_WT_PARAM_MAX = new Double(Double.MAX_VALUE);
-	private final static Double REL_A_PRIORI_WT_PARAM_DEFAULT = new Double(100);
+	private final static Double REL_A_PRIORI_WT_PARAM_DEFAULT = new Double(1);
 	private final static String REL_A_PRIORI_WT_PARAM_INFO = "Relative to that put on the sement slip rates";
 	private DoubleParameter relativeA_PrioriWeightParam;
